+++ src/server.ts
@@ -2,8 +2,9 @@ import express from "express";
 import type compression from "compression";
 import http from "node:http";
 import https from "node:https";
+import { BuiltinLogger } from "./builtin-logger";
 import { AppConfig, CommonConfig, ServerConfig } from "./config-type";
-import { createLogger, isActualLogger } from "./logger";
+import { isLoggerInstance } from "./logger-helpers";
 import { loadPeer } from "./peer-helpers";
 import { defaultResultHandler } from "./result-handler";
 import { Parsers, Routing, initRouting } from "./routing";
@@ -21,9 +22,9 @@ const makeCommonEntities = (config: CommonConfig) => {
     console.log(getStartupLogo());
   }
   const errorHandler = config.errorHandler || defaultResultHandler;
-  const rootLogger = isActualLogger(config.logger)
+  const rootLogger = isLoggerInstance(config.logger)
     ? config.logger
-    : createLogger(config.logger);
+    : new BuiltinLogger(config.logger);
   rootLogger.debug("Running", process.env.TSUP_BUILD || "from sources");
   const loggingMiddleware = createLoggingMiddleware({ rootLogger, config });
   const notFoundHandler = createNotFoundHandler({ rootLogger, errorHandler });