+++ tests/unit/logger-helpers.spec.ts
@@ -1,0 +1,34 @@
+import { BuiltinLogger } from "../../src";
+import { BuiltinLoggerConfig } from "../../src/builtin-logger";
+import { AbstractLogger, isLoggerInstance } from "../../src/logger-helpers";
+import { describe, expect, test } from "vitest";
+
+describe("Logger helpers", () => {
+  describe("isLoggerInstance()", () => {
+    test.each<BuiltinLoggerConfig>([
+      { level: "silent" },
+      { level: "debug", color: false },
+      { level: "info", color: true },
+      { level: "warn", depth: 5 },
+      { level: "warn", depth: null },
+      { level: "warn", depth: Infinity },
+    ])("should invalidate built-in logger config %#", (sample) => {
+      expect(isLoggerInstance(sample)).toBeFalsy();
+    });
+
+    test.each<AbstractLogger>([
+      // issue #1605: should not allow methods
+      { level: "debug", debug: () => {} },
+      { level: "warn", error: () => {} },
+      // issue #1772: similar to #1605, but the methods are in prototype
+      new (class {
+        level = "debug";
+        debug() {}
+      })(),
+      Object.setPrototypeOf({ level: "debug" }, { debug: () => {} }),
+      new BuiltinLogger({ level: "debug" }),
+    ])("should validate logger instances %#", (sample) => {
+      expect(isLoggerInstance(sample)).toBeTruthy();
+    });
+  });
+});